Move model from internal to pkg
