+++ internal/api/handlers/v0/auth/github_oidc.go
@@ -15,7 +15,7 @@ import (
 	v0 "github.com/modelcontextprotocol/registry/internal/api/handlers/v0"
 	"github.com/modelcontextprotocol/registry/internal/auth"
 	"github.com/modelcontextprotocol/registry/internal/config"
-	"github.com/modelcontextprotocol/registry/internal/model"
+	"github.com/modelcontextprotocol/registry/pkg/model"
 )
 
 // GitHubOIDCTokenExchangeInput represents the input for GitHub OIDC token exchange
